fixed Database.CurrentSize when over 8gb (2gb max int << 2) (had int instead of long)

+++ src/main/java/suneido/database/server/DbmsRemote.java
@@ -271,7 +271,7 @@ public class DbmsRemote extends Dbms {
 	private static long getLong(Scanner scan, char type) {
 		try {
 			scan.skip("\\s*" + type);
-			return scan.nextInt();
+			return scan.nextLong();
 		} catch (NoSuchElementException e) {
 			return ERR;
 		}
